% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_core_data}
\alias{load_core_data}
\title{Load and Prepare Core Datasets}
\usage{
load_core_data()
}
\value{
A tibble with the combined data.
}
\description{
Reads and joins the core olist datasets from within the package.
}
