% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_sales_data}
\alias{clean_sales_data}
\title{Clean and Transform Sales Data}
\usage{
clean_sales_data(data)
}
\value{
A cleaned and transformed dataframe.
}
\description{
Takes the raw, combined olist data and performs cleaning.
}
